Migrated the functionality of the forms to update users and to create projects to new pages. Made various minor changes to multiple page components and added the lean option in various queries in order to avoid exceeding the stack when making requests.
