+++ client/src/pages/Sign-Up.jsx
@@ -7,6 +7,14 @@ import { ADD_USER } from '../utils/mutations';
 import Auth from '../utils/auth';
 
 const SignUp = () => {
+
+  useEffect(() => {
+    document.title = `Register to start a campaign/project!`
+    return () => {
+      if (location.pathname !== '/signup') document.title = 'Unity Fund'
+    }
+  }, [])
+
   const initialState = {
     username: '',
     email: '',
