+++ client/src/pages/SingleProject.jsx
@@ -7,12 +7,16 @@ function SingleProject({ projectName, projectDate ='12/6/23.', goalAmount, expir
         <>
             <Card >
                 <Card.Header>
-                    {projectName} {goalAmount} {expiresIn}
+                    <h3>{projectName}</h3>
+                    We need your help to raise ${goalAmount}!<br/>
+                    Hurry! This campaign expires in {expiresIn} days!
                 </Card.Header>
                 <Card.Title>{userId}</Card.Title>
-                <Card.Subtitle> {projectDate}</Card.Subtitle>
+                <Card.Subtitle> This project was submitted on {projectDate}</Card.Subtitle>
+                <hr/>
                 <Card.Body>
                 <Image src={projectImage}/>
+                <h6>Description</h6>
                   {projectDescription}
                 </Card.Body>
                 <Card.Footer>
